+++ packages/flopper/test/finiteShifter.test.js
@@ -1,0 +1,9 @@
+import { init }         from '@vect/vector-init'
+import { test }          from 'node:test'
+import { finiteShifter } from '../src/utils/finiteShifter.js'
+
+test('shiftFlopper', () => {
+  const vec = init(24, i => i)
+  console.log([ ...finiteShifter(vec, 1) ])
+  console.log([ ...finiteShifter(vec, 4) ])
+})