+++ packages/munsell/src/extends/RGB.js
@@ -1,7 +1,7 @@
 import { hexToRgb, hslToRgb } from '@palett/convert'
 import { minEntry }           from '../../utils/minBy.js'
 import { Cova }          from '../Cova.js'
-import { Domain }        from '../../resources/Domain.js'
+import { Domain }        from '../../archive/Domain.js'
 import { RGB as RawRGB } from '@palett/color-space'
 
 export class RGB extends RawRGB {