+++ packages/fluo-entries/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @palett/fluo-entries
 
+## 0.12.12
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/fluo@0.12.12
+
 ## 0.12.11
 
 ### Patch Changes